pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yuichatbox"
version = "1.0.0"
description = "AI Chat Application with Multi-Model Support and Reasoning Visualization"
authors = [{name = "YUI ChatBox Team"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["chatbot", "ai", "openai", "llm", "chat", "streaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "openai==1.10.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.3",
    "pydantic-settings>=2.0.0",
    "httpx==0.26.0",
    "click>=8.0.0",
    "sqlalchemy>=2.0.0",
    "nanoid>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
yui = "yuichatbox.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/yui-chatbox"
Documentation = "https://github.com/yourusername/yui-chatbox#readme"
Repository = "https://github.com/yourusername/yui-chatbox"
Issues = "https://github.com/yourusername/yui-chatbox/issues"

[tool.setuptools]
packages = ["yuichatbox"]

[tool.setuptools.package-data]
yuichatbox = ["static/**/*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"
